+++ src/ui/qt/operate_home.py
@@ -22,6 +22,7 @@ from PyQt5.QtWidgets import (
 from src.ui.qt.ble_domain import BleDomainView
 from src.ui.qt.domain_grid import DomainGrid
 from src.ui.qt.gps_domain import GpsDomainView
+from src.ui.qt.subghz_domain import SubGhzDomainView
 from src.ui.qt.wifi_domain import WifiDomainView
 
 
@@ -45,6 +46,7 @@ class OperateHome(QWidget):
     def __init__(self, wifi_center: "Optional[QWidget]" = None,
                  ble_center: "Optional[QWidget]" = None,
                  gps_center: "Optional[QWidget]" = None,
+                 subghz_center: "Optional[QWidget]" = None,
                  parent: "Optional[QWidget]" = None) -> None:
         super().__init__(parent)
         self._current: Optional[str] = None
@@ -63,6 +65,8 @@ class OperateHome(QWidget):
                 view = BleDomainView(center=ble_center or QWidget())
             elif key == "gps":
                 view = GpsDomainView(center=gps_center or QWidget())
+            elif key == "subghz":
+                view = SubGhzDomainView(center=subghz_center or QWidget())
             else:
                 view = _placeholder(titles.get(key, key))
             self._views[key] = view